ID-39: Refuse to delete payment method that does not belong to requesting user

+++ src/Application/Actions/DeletePaymentMethod.php
@@ -12,6 +12,7 @@ use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Log\LoggerInterface;
 use Stripe\Exception\ApiErrorException;
 use Stripe\Exception\InvalidRequestException;
+use Stripe\PaymentMethod;
 use Stripe\StripeClient;
 
 class DeletePaymentMethod extends Action
@@ -35,6 +36,28 @@ class DeletePaymentMethod extends Action
 
         $paymentMethodId = $args['payment_method_id'];
         \assert(is_string($paymentMethodId));
+
+        $allPaymentMethodsOfCustomer = $this->stripeClient->customers->allPaymentMethods(
+            $customerId,
+        )->toArray();
+
+        $allCustomersPaymentMethodIds = array_map(
+            static fn(PaymentMethod $pm) => $pm->id,
+            $allPaymentMethodsOfCustomer
+        );
+
+        if (!in_array($paymentMethodId, $allCustomersPaymentMethodIds, true)) {
+            $this->logger->warning(
+                "Refusing to delete stripe payment method as not found for customer",
+                compact('customerId', 'paymentMethodId')
+            );
+
+            return $this->respondWithData(
+                $response,
+                ['error' => 'Payment method not found'],
+                StatusCodeInterface::STATUS_BAD_REQUEST
+            );
+        }
 
         try {
             $this->stripeClient->paymentMethods->detach($paymentMethodId);
